// lib/api/generated/models/planBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */
import type { PlanAssignment } from './planAssignment';
import type { PlanOrigin } from './planOrigin';
import type { PlanStatus } from './planStatus';
import type { PlanWeek } from './planWeek';

export interface PlanBody {
  assignment?: PlanAssignment;
  is_current?: boolean;
  /** @nullable */
  notes?: string | null;
  origin?: null | PlanOrigin;
  /** @nullable */
  starts_on?: string | null;
  status?: PlanStatus;
  title: string;
  weeks?: PlanWeek[];
}