// lib/api/generated/models/planStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type PlanStatus = typeof PlanStatus[keyof typeof PlanStatus];


export const PlanStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
  catalog: 'catalog',
} as const;